+++ src/replication-protocol/upstream.ts
@@ -16,7 +16,8 @@ import {
     batchArray,
     ensureNotFalsy,
     parseRevision,
-    PROMISE_RESOLVE_FALSE
+    PROMISE_RESOLVE_FALSE,
+    randomCouchString
 } from '../plugins/utils';
 import {
     getLastCheckpointDoc,
@@ -195,6 +196,7 @@ export function startReplicationUpstream<RxDocType, CheckpointType>(
         docs: RxDocumentData<RxDocType>[],
         checkpoint: CheckpointType
     ): Promise<boolean> {
+        const callId = randomCouchString(10);
         state.stats.up.persistToMaster = state.stats.up.persistToMaster + 1;
 
         /**
@@ -300,7 +302,7 @@ export function startReplicationUpstream<RxDocType, CheckpointType>(
             const writeBatches = batchArray(writeRowsArray, state.input.pushBatchSize);
             await Promise.all(
                 writeBatches.map(async (writeBatch) => {
-                    const masterWriteResult = await replicationHandler.masterWrite(writeBatch);
+                    const masterWriteResult = await replicationHandler.masterWrite(writeBatch, { callId });
                     masterWriteResult.forEach(conflictDoc => {
                         const id = (conflictDoc as any)[state.primaryPath];
                         conflictIds.add(id);
@@ -323,7 +325,7 @@ export function startReplicationUpstream<RxDocType, CheckpointType>(
             if (useWriteRowsToMeta.length > 0) {
                 await state.input.metaInstance.bulkWrite(
                     useWriteRowsToMeta,
-                    'replication-up-write-meta'
+                    'replication-up-write-meta-' + callId
                 );
                 // TODO what happens when we have conflicts here?
             }